backend/requirements.txt
python-dotenv
Flask
flask-cors
Flask-JWT-Extended
Flask-Migrate
Flask-SQLAlchemy
Werkzeug
requests
openai
torch
transformers
huggingface-hub
tokenizers
safetensors
langdetect
PyPDF2
deep_translator
gtts
psycopg2-binary
pytesseract
SpeechRecognition
pydub
logging
streamlit-realtime-audio-recorder
pytesseract
gunicorn